requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
mssql-django>=1.3,<2.0
django-cors-headers>=4.3,<5.0
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.5,<3.0
google-api-python-client>=2.100,<3.0
google-auth-oauthlib>=1.1,<2.0
google-auth-httplib2>=0.1,<1.0
python-decouple>=3.8,<4.0
pyodbc>=5.0,<6.0
django-redis>=5.4,<6.0
